Palestrantes e Redes Sociais Frontend Começo

+++ Backend/src/ProEventos.API/Helpers/Util.cs
@@ -37,9 +37,12 @@ namespace ProEventos.API.Helpers
 
         public void DeleteImage(string imageName, string destino)
         {
+            if (!string.IsNullOrEmpty(imageName)) 
+            {
             var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @$"Resources/{destino}",
             imageName);
             if (System.IO.File.Exists(imagePath)) System.IO.File.Delete(imagePath);
+            }
         }
     }
 }
